backend/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
fastapi==0.128.0
h11==0.16.0
idna==3.11
PyPDF2==3.0.1
python-dotenv==1.2.1
python-multipart==0.0.21
requests==2.32.5
SQLAlchemy==2.0.23
psycopg[binary]==3.3.2
openai==1.58.1
starlette==0.50.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
